Initial Commit
This project entailed creating the full logic for a classic pinball
arcade machine. The machine had been retrofitted with an Arduino Mega
2560. I worked with one other person on this project, and while there
was no clear division of labor, the both of us were present and
contributing at all times when working on the project. We were supplied
with a C++ library to ease the usage of the physical components of the
machine, though we had done this by hand in previous projects. Other
than in using the provided library, we were instructed to write C style
code. The program could accept up to 4 players, with 3 balls each.
Lights would indicate the ball number, current player, current bonuses,
and current score. At the end, the high score would be displayed, and
kept until the machine was turned off.

I cannot find a copy of the BallyLib, so this will not compile